Remove worker queue adapters and skynet_worker_queue Last commit missed the following message:
Skynet Workers now communicate with skynet manager by grabbing the manager over DRB and passing hashes directly to an internal
Queue object held by the manager.
The manager has a thread that watches the queue and updates the @worker_queue when new items are in the queue.
The worker/manager no longer use tuplespace to communicate

Removed unecessary logging in a few places.